+++ changelog.md
@@ -2,5 +2,5 @@
 
 ### [Push](https://github.com/emartech/android-emarsys-sdk/wiki#2-push)
 
-* Fixed an issue when both Huawei and Firebase were used as Push providers, push messages might have
-  failed to arrive.
+* Fixed an issue what could cause opening the host Application from a Push Notification fail on
+  Android 12 and 13.